requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-admin-thumbnails==0.2.8
django-bootstrap4==23.2
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5
mysqlclient==2.0.3
notification==0.2.1
Pillow==10.0.1
pip==23.2.1
pytz==2023.3.post1
setuptools==68.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
